# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HapiControl Internal Financial Operations API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # CLIENTS
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments
      parameters:
        - name: clientId
          in: query
          schema:
            type: ["integer", "null"]
        - name: creditId
          in: query
          schema:
            type: ["integer", "null"]
        - name: executiveId
          in: query
          schema:
            type: ["integer", "null"]
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Payments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Register payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get payment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # PAYMENT COMMITMENTS
...
  /public/requests:
    post:
      operationId: createPublicRequest
      tags: [public]
      summary: Submit public information request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePublicRequestBody"
      responses:
        "201":
          description: Request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, executive, client, public]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, username, fullName, role, isActive, createdAt, updatedAt]
    CreateUserBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        fullName:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, executive, client, public]
      required: [username, password, fullName, role]
    UpdateUserBody:
      type: object
      properties:
        fullName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
        password:
          type: ["string", "null"]
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        creditId:
          type: integer
        paymentNumber:
          type: integer
        paymentDate:
          type: string
        amountPaid:
          type: number
        amountExpected:
          type: number
        lateFee:
          type: ["number", "null"]
        updatedBalance:
          type: number
        paymentStatus:
          type: string
          enum: [on_time, partial, late, missed]
        executiveId:
          type: ["integer", "null"]
        executiveName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, clientId, creditId, paymentNumber, paymentDate, amountPaid, amountExpected, updatedBalance, paymentStatus, createdAt]
    CreatePaymentBody:
      type: object
      properties:
        clientId:
          type: integer
        creditId:
          type: integer
        paymentDate:
          type: string
        amountPaid:
          type: number
        lateFee:
          type: ["number", "null"]
        executiveId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
      required: [clientId, creditId, paymentDate, amountPaid]
...
    CreatePublicRequestBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        message:
          type: string
      required: [name, phone, message]